+++ src/tasmotaSwitchService.ts
@@ -44,6 +44,7 @@ export class tasmotaSwitchService {
       accessory.context.mqttHost.on(accessory.context.device[this.uniq_id].stat_t, this.statusUpdate.bind(this));
       accessory.context.mqttHost.on(accessory.context.device[this.uniq_id].avty_t, this.availabilityUpdate.bind(this));
     }
+    nunjucks.installJinjaCompat();
     nunjucks.configure({
       autoescape: true,
     });